Moved extract modules code to cb class
